Fix major gsbase bug, implement and fix clock_gettime (API has 68 year uptime limit), fix user vfprintf, fix truncated copy from user helper, remove some PCI debug spam

+++ kernel/syscall/sys_time.cc
@@ -42,8 +42,8 @@ int sys_clock_gettime(clockid_t id, timespec *tm)
     uint64_t t = time_ns();
 
     timespec result{};
-    result.tv_sec = t / 1000000000;
-    result.tv_nsec = t % 1000000000;
+    result.tv_sec = int32_t(t / 1000000000);
+    result.tv_nsec = int32_t(t % 1000000000);
 
     return clock_copy_timespec_to_user(tm, result);
 }
